// src/components/ArchivesList/ArchivesList.types.ts
export interface ArchiveItem {
  audio_length: number
  audio_quality: string
  date: string
  description: string
  end_time?: any
  id: number
  mixcloud_id: string
  name: string
  picture_url: string
  preview_url: string
  primary_color: string
  publish_date: Date
  slug: string
  start_time?: any
  tags?: string
  genres?: Array<{ id: number; name: string }>
}